+++ .env.goerli
@@ -1,0 +1,2 @@
+NETWORK='Goerli'
+RPC_URL='https://goerli.infura.io/v3/9aa3d95b3bc440fa88ea12eaa4456161'
