+++ Cargo.toml
@@ -14,3 +14,4 @@ tokio = "1.17.0"
 chrono = "0.4.19"
 dotenv = "0.15.0"
 colored = "2.0.0"
+bs58 = "0.4.0"
