+++ src/errors.rs
@@ -1,4 +1,3 @@
-
 use std::fmt::Display;
 use colored::Colorize;
 
@@ -6,7 +5,7 @@ use colored::Colorize;
 pub enum MemoParserError {
     ParseError(String),
     FetchError(String),
-    InternalError(String),
+    //InternalError(String),
 }
 
 impl Display for MemoParserError {
@@ -14,7 +13,7 @@ impl Display for MemoParserError {
         match self {
             MemoParserError::ParseError(message) => write!(f, "{}: {}", "cannot parse calldata".red(), message),
             MemoParserError::FetchError(message) => write!(f, "{}: {}", "cannot fetch calldata".red(), message),
-            MemoParserError::InternalError(message) => write!(f, "{}: {}", "internal error".red(), message),
+            //MemoParserError::InternalError(message) => write!(f, "{}: {}", "internal error".red(), message),
         }
     }
 }
