+++ src/lib.rs
@@ -1,4 +1,4 @@
 pub mod ethutils;
-pub mod memo;
-pub mod memoparser;
+pub mod utils;
 pub mod errors;
+pub mod calldata;
